Better documentation

+++ server/src/config.rs
@@ -1,6 +1,9 @@
 use bevy::prelude::Resource;
 
+/// Various options about how Stardust should manage the server.
+/// You can change this at any time, even when the server is running.
 #[derive(Resource, Debug, Clone)]
 pub struct ServerConfig {
-    pub max_players: u32,
+    /// The soft max amount of players that can be connected at any one time.
+    pub max_players: u16,
 }
